rename src to edgar_tool
