+++ edgar_tool/constants.py
