+++ edgar_tool/rss.py
@@ -7,9 +7,9 @@ import requests
 import xmltodict
 from requests import Response
 
-from src.constants import RSS_FEED_CSV_FIELDS_NAMES
-from src.io import write_results_to_file
-from src.utils import safe_get, unpack_singleton_list
+from edgar_tool.constants import RSS_FEED_CSV_FIELDS_NAMES
+from edgar_tool.io import write_results_to_file
+from edgar_tool.utils import safe_get, unpack_singleton_list
 
 RSS_FEED_DATA_DIRECTORY = Path(__file__).resolve().parents[1] / "data"
 RSS_FEED_URL = "https://www.sec.gov/Archives/edgar/xbrlrss.all.xml"
