+++ edgar_tool/text_search.py
@@ -9,22 +9,22 @@ from bs4 import BeautifulSoup
 from selenium.webdriver.common.by import By
 from selenium.webdriver.remote.webelement import WebElement
 
-from src.browser import (
+from edgar_tool.browser import (
     BrowserDriver,
     fetch_page,
     extract_html_table_rows,
     PageCheckFailedError,
     ResultsTableNotFoundError,
 )
-from src.constants import (
+from edgar_tool.constants import (
     TEXT_SEARCH_BASE_URL,
     TEXT_SEARCH_FILING_CATEGORIES_MAPPING,
     TEXT_SEARCH_RESULTS_TABLE_XPATH,
     TEXT_SEARCH_SPLIT_BATCHES_NUMBER,
     TEXT_SEARCH_CSV_FIELDS_NAMES,
 )
-from src.io import write_results_to_file
-from src.utils import split_date_range_in_n, unpack_singleton_list
+from edgar_tool.io import write_results_to_file
+from edgar_tool.utils import split_date_range_in_n, unpack_singleton_list
 
 
 class EdgarTextSearcher:
